underhill_mem: Separate out modules
